requirements.txt
openai==1.12.0
python-dotenv==1.0.0
flask==3.0.2
httpx==0.27.0
